Fixes after review, removal of plant icons, deleting files along with plant deletion

+++ PlantManager/src/main/java/com/ajaskiewicz/PlantManager/web/rest/PlantController.java
@@ -1,10 +1,8 @@
 package com.ajaskiewicz.PlantManager.web.rest;
 
 import com.ajaskiewicz.PlantManager.model.Plant;
-import com.ajaskiewicz.PlantManager.service.PlantService;
-import com.ajaskiewicz.PlantManager.service.RoomService;
-import com.ajaskiewicz.PlantManager.service.UserService;
-import com.ajaskiewicz.PlantManager.service.WateringScheduleService;
+import com.ajaskiewicz.PlantManager.service.*;
+import com.ajaskiewicz.PlantManager.web.utils.FileDeleteUtil;
 import com.ajaskiewicz.PlantManager.web.utils.FileUploadUtil;
 import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,56 +18,71 @@ import java.util.Optional;
 @RequestMapping(value = "/dashboard")
 public class PlantController {
 
-    @Autowired
     private PlantService plantService;
-
-    @Autowired
     private RoomService roomService;
-
-    @Autowired
     private WateringScheduleService wateringScheduleService;
+    private UserService userService;
+    private SecurityService securityService;
 
     @Autowired
-    private UserService userService;
+    public PlantController(PlantService plantService, RoomService roomService, WateringScheduleService wateringScheduleService, UserService userService, SecurityService securityService) {
+        this.plantService = plantService;
+        this.roomService = roomService;
+        this.wateringScheduleService = wateringScheduleService;
+        this.userService = userService;
+        this.securityService = securityService;
+    }
 
     @RequestMapping
     public String showDashboard(Model model) {
+        if (!securityService.isAuthenticated()) { return "redirect:/"; }
+
         model.addAttribute("plant", plantService.findAllByUserId(userService.findIdOfLoggedUser()));
         return "dashboardPage";
     }
 
     @RequestMapping(path = {"/editPlant", "/editPlant/{id}"})
-    public String editPlantById(Model model, @PathVariable("id") Optional<Integer> id)
-            throws NotFoundException {
+    public String editPlantById(Model model, @PathVariable("id") Optional<Integer> id) throws NotFoundException {
+        if (!securityService.isAuthenticated()) { return "redirect:/"; }
+
         if (id.isPresent()) {
-            Plant entity = plantService.find(id.get());
+            var entity = plantService.find(id.get());
             model.addAttribute("plant", entity);
         } else {
             model.addAttribute("plant", new Plant());
         }
+
         return "editPage";
     }
 
     @RequestMapping(path = "/addPlant", method = RequestMethod.POST)
     public String createOrUpdatePlant(Plant plant, @RequestParam("image") MultipartFile multipartFile) throws IOException {
+        if (!securityService.isAuthenticated()) { return "redirect:/"; }
 
-        String filename = multipartFile.getOriginalFilename();
+        var filename = multipartFile.getOriginalFilename();
 
         if (multipartFile.isEmpty()) {
             plantService.createOrUpdatePlant(plant);
         } else {
             plant.setImageName(filename);
-            Plant savedPlant = plantService.createOrUpdatePlant(plant);
+            var savedPlant = plantService.createOrUpdatePlant(plant);
 
-            String uploadDirectory = "uploadedImages/" + savedPlant.getId();
+            var uploadDirectory = "uploadedImages/" + savedPlant.getId();
             FileUploadUtil.saveFile(uploadDirectory, filename, multipartFile);
         }
+
         return "redirect:/dashboard";
     }
 
     @RequestMapping(value = "/deletePlant/{id}")
-    public String deletePlantById(@PathVariable("id") Integer id) throws NotFoundException {
+    public String deletePlantById(@PathVariable("id") Integer id) throws NotFoundException, IOException {
+        if (!securityService.isAuthenticated()) { return "redirect:/"; }
+
+        var deleteDirectory = "uploadedImages/" + id;
+        FileDeleteUtil.deleteFile(deleteDirectory);
+
         plantService.delete(id);
+
         return "redirect:/dashboard";
     }
